fix: Solve post merge conflicts
* It was necessary to move some constants to constants.py
  and modify related code accordingly
* Small fixes and refactoring in other files
